VBox additions - Add custom README file. - Don't output pszParm for flag != FSA_ATTACH, avoiding a trap when FC/2 gives us a buffer with a file contents, after we opened some file for writing, then closed it. So, the same buffer with the file contents are given as a pszParm FS_ATTACH parameter (to query UNC resource name). - Convert DOS-style wildcards to WinNT-style ones (like Knut does in his vboxsf.ifs version). But contrary to Knut's solution, we added this conversion to FS_PROCCESSNAME. Our solution is a bit shorter. - Pack binaries with lxlite.
git-svn-id: http://svn.netlabs.org/repos/vbox/trunk@192 8b531da7-8337-e511-8bd3-62de69f6b5d0

+++ src/VBox/Additions/os2/VBoxFS/VBoxFSStr.cpp
@@ -225,7 +225,7 @@ int tolower (int c)
     return c;
 }
 
-char *strncpy(char *dst, char *src, int len)
+char *strncpy(char *dst, char *src, size_t len)
 {
     char *ret;
